+++ src/query/phrase_prefix_query/phrase_prefix_query.rs
@@ -111,8 +111,16 @@ impl PhrasePrefixQuery {
         }
         let terms = self.phrase_terms();
         let bm25_weight_opt = match enable_scoring {
-            EnableScoring::Enabled { searcher, .. } => {
-                Some(Bm25Weight::for_terms(searcher, &terms)?)
+            EnableScoring::Enabled {
+                searcher,
+                statistics_provider,
+            } => {
+                let bm25_params = searcher.bm25_params();
+                Some(Bm25Weight::for_terms(
+                    statistics_provider,
+                    &terms,
+                    bm25_params,
+                )?)
             }
             EnableScoring::Disabled { .. } => None,
         };